# Skip stackmaps for yk_trace_basicblock.