Related: https://github.com/ykjit/yk/pull/939
Source: https://github.com/ykjit/yk/pull/939#issuecomment-1909980698

The yk_trace_basicblock call doesn't require stackmaps therefore we can skip stackmap adjustment for those functions.

Tests:
```
Testing Time: 405.04s
  Skipped          :    46
  Unsupported      :   885
  Passed           : 92778
  Expectedly Failed:   211
```