@ltratt 2024-02-13
Don't we need to guard this to only when we're compiling in swt mode? Also, what performance difference does this make (roughly speaking)?