@Pavel-Durov 2024-02-13
> Don't we need to guard this to only when we're compiling in swt mode? 

Yes, I'll add the guard.

> Also, what performance difference does this make (roughly speaking)?

Performance difference between `swt` and `hwt`?