@ltratt 2024-02-13
>>    Also, what performance difference does this make (roughly speaking)?
> Performance difference between swt and hwt?

The performance difference of not adding stackmaps to yk_trace_basicblock vs. adding it.