@Pavel-Durov 2024-02-20
> Don't we need to guard this to only when we're compiling in swt mode? Also, what performance difference does this make (roughly speaking)?

@ltratt LLVM passes are compile time so the impact of adding a guard here will only affect compilation time which is insignificant for the runtime performance impact.