@Pavel-Durov 2024-02-22
## Benchmarks
### With yk_trace_basicblock stackmaps
```
hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/closure.lua' 

Benchmark 1: ./src/lua ./tests/closure.lua
  Time (mean ± σ):      4.439 s ±  0.411 s    [User: 4.569 s, System: 0.186 s]
  Range (min … max):    3.597 s …  5.049 s    30 runs

hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/strings.lua' 
Benchmark 1: ./src/lua ./tests/strings.lua
  Time (mean ± σ):      3.481 s ±  0.351 s    [User: 3.521 s, System: 0.097 s]
  Range (min … max):    3.245 s …  4.120 s    30 runs
```
### Without yk_trace_basicblock stackmaps

```
hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/closure.lua'
Benchmark 1: ./src/lua ./tests/closure.lua
  Time (mean ± σ):      4.022 s ±  0.464 s    [User: 4.020 s, System: 0.127 s]
  Range (min … max):    3.318 s …  4.736 s    30 runs

hyphyperfine --min-runs 30 './src/lua ./tests/strings.lua' 
Benchmark 1: ./src/lua ./tests/strings.lua
  Time (mean ± σ):      3.506 s ±  0.441 s    [User: 3.537 s, System: 0.096 s]
  Range (min … max):    2.649 s …  4.116 s    30 runs
```