@ltratt 2024-02-22
We're still getting seriously noisy numbers here, so it's very hard to conclude much :/ The range for `strings.lua` is "2.649s …  4.116s" which I'm really struggling to come up with an explanation for.