@Pavel-Durov 2024-02-22
Ran again multiple times on a completely quite machine (no anomalies were reported by hyperfine)
The variation is not as big as before for `strings.lua`.

## Benchmarks

### With yk_trace_basicblock stackmaps
```
hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/closure.lua' 
Benchmark 1: ./src/lua ./tests/closure.lua
  Time (mean ± σ):      4.112 s ±  0.097 s    [User: 4.139 s, System: 0.064 s]
  Range (min … max):    3.794 s …  4.325 s    30 runs

hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/strings.lua' 
Benchmark 1: ./src/lua ./tests/strings.lua
  Time (mean ± σ):      3.896 s ±  0.098 s    [User: 3.937 s, System: 0.051 s]
  Range (min … max):    3.711 s …  4.069 s    30 runs
```
### Without yk_trace_basicblock stackmaps

```
hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/closure.lua' 
Benchmark 1: ./src/lua ./tests/closure.lua
  Time (mean ± σ):      3.503 s ±  0.069 s    [User: 3.520 s, System: 0.060 s]
  Range (min … max):    3.317 s …  3.624 s    30 runs
 
hyperfine --warmup 10 --min-runs 30 './src/lua ./tests/strings.lua' 
Benchmark 1: ./src/lua ./tests/strings.lua
  Time (mean ± σ):      3.248 s ±  0.070 s    [User: 3.280 s, System: 0.045 s]
  Range (min … max):    3.112 s …  3.371 s    30 runs
 ```