@ltratt 2024-02-22
Ok those look better! So roughly a 10% difference from removing stackmaps here. Very interesting -- thank you for this!

[If you're wondering why I was wondering, that means I can now guesstimate the cost of stackmaps on performance of the rest of the interpreter as ~5-7%. Not great, but not quite as bad as I feared.]